backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9
python-slugify==8.0.4
email-validator==2.1.1
